requirements/local.txt
# Local development dependencies go here
-r base.txt
bpython==0.14
django-debug-toolbar==1.2.2
Sphinx==1.2.3
Werkzeug==0.10.1
